refactor: move utility methods to dedicated file
